Module iota_stronghold::actors::client

Enums

ProcResult
A Procedure return result type. Contains the different return values for the Procedure type calls used with Stronghold.runtime_exec(...).

Procedure
Procedure type used to call to the runtime via Strongnhold.runtime_exec(...).

SHRequest

SHResults
Messages that come from stronghold

SLIP10DeriveInput
SLIP10DeriveInput type used to specify a Seed location or a Key location for the SLIP10Derive procedure.

SerdeProcResult