Enum iota_stronghold::actors::internal::InternalMsg

pub enum InternalMsg {

    CreateVault(VaultId, RecordId),
    WriteToVault(VaultId, RecordId, Vec<u8>, RecordHint),
    RevokeData(VaultId, RecordId),
    GarbageCollect(VaultId),
    ListIds(VaultId),
    CheckRecord(VaultId, RecordId),
    ReadSnapshot(Key, Option<String>, Option<PathBuf>, ClientId, Option<ClientId>),
    ReloadData {
        id: ClientId,
        data: Box<(HashMap<VaultId, Key<Provider>>, DbView<Provider>, Cache<Vec<u8>, Vec<u8>>)>,
        status: StatusMessage,
    },
    ClearCache,
    KillInternal,
    FillSnapshot {
        client: Client,
    },
    SLIP10Generate {
        vault_id: VaultId,
        record_id: RecordId,
        hint: RecordHint,
        size_bytes: usize,
    },
    SLIP10DeriveFromSeed {
        chain: Chain,
        seed_vault_id: VaultId,
        seed_record_id: RecordId,
        key_vault_id: VaultId,
        key_record_id: RecordId,
        hint: RecordHint,
    },
    SLIP10DeriveFromKey {
        chain: Chain,
        parent_vault_id: VaultId,
        parent_record_id: RecordId,
        child_vault_id: VaultId,
        child_record_id: RecordId,
        hint: RecordHint,
    },
    BIP39Generate {
        passphrase: String,
        vault_id: VaultId,
        record_id: RecordId,
        hint: RecordHint,
    },
    BIP39Recover {
        mnemonic: String,
        passphrase: String,
        vault_id: VaultId,
        record_id: RecordId,
        hint: RecordHint,
    },
    Ed25519PublicKey {
        vault_id: VaultId,
        record_id: RecordId,
    },
    Ed25519Sign {
        vault_id: VaultId,
        record_id: RecordId,
        msg: Vec<u8>,
    },

}

Messages used for the Internal Actor.

Variants

CreateVault(VaultId, RecordId)

Creates a new vault at the given VaultId and RecordId

WriteToVault(VaultId, RecordId, Vec<u8>, RecordHint)

Writes data to a vault at the location of the given VaultId and RecordId

RevokeData(VaultId, RecordId)

Revokes data from a vault at the location of the given VaultId and RecordId

GarbageCollect(VaultId)

Garbage collects a vault at the given VaultId

ListIds(VaultId)

Lists ids of the vault at the given VaultId

CheckRecord(VaultId, RecordId)

Checks to see if a record exists at the given VaultId and RecordId

ReadSnapshot(Key, Option<String>, Option<PathBuf>, ClientId, Option<ClientId>)

Reads the snapshot from the file.

ReloadData

Reloads the data from the snapshot.

Fields of ReloadData

id: ClientIddata: Box<(HashMap<VaultId, Key<Provider>>, DbView<Provider>, Cache<Vec<u8>, Vec<u8>>)>status: StatusMessage

ClearCache

Clears the cache.

KillInternal

Kills the internal actor.

FillSnapshot

Fills the snapshot state with the current internal state.

Fields of FillSnapshot

client: Client

SLIP10Generate

[SLIP10Generate] seed Proc.

Fields of SLIP10Generate

vault_id: VaultIdrecord_id: RecordIdhint: RecordHintsize_bytes: usize

SLIP10DeriveFromSeed

[SLIP10DeriveFromSeed] Proc

Fields of SLIP10DeriveFromSeed

chain: Chainseed_vault_id: VaultIdseed_record_id: RecordIdkey_vault_id: VaultIdkey_record_id: RecordIdhint: RecordHint

SLIP10DeriveFromKey

[SLIP10DeriveFromKey] Proc

Fields of SLIP10DeriveFromKey

chain: Chainparent_vault_id: VaultIdparent_record_id: RecordIdchild_vault_id: VaultIdchild_record_id: RecordIdhint: RecordHint

BIP39Generate

[BIP39Generate] Proc

Fields of BIP39Generate

passphrase: Stringvault_id: VaultIdrecord_id: RecordIdhint: RecordHint

BIP39Recover

[BIP39Recover] Proc

Fields of BIP39Recover

mnemonic: Stringpassphrase: Stringvault_id: VaultIdrecord_id: RecordIdhint: RecordHint

Ed25519PublicKey

[Ed25519PublicKey] Proc

Fields of Ed25519PublicKey

vault_id: VaultIdrecord_id: RecordId

Ed25519Sign

[Ed25519Sign] Proc

Fields of Ed25519Sign

vault_id: VaultIdrecord_id: RecordIdmsg: Vec<u8>