Module iota_stronghold::actors::registry

Registry Actor

The registry actor keeps record of all client actors bound to a unique client_id. SecureClient actors can be added, removed or queried for their [actix::Addr]. The registry can also be queried for the snapshot actor.

Modules

messages

Structs

Registry
Registry [Actor], that owns [Client] actors, and manages them. The registry can be modified

Enums

RegistryError