Module iota_stronghold::actors::secure

Secure Actor module

The secure actor runs as service, isolates contained data, and synchronizes across multiple accesses.

Re-exports

pub use crate::actors::GetSnapshot;

pub use crate::actors::Registry;

pub use crate::internals::Provider;

pub use crate::state::key_store::KeyStore;

pub use crate::state::secure::SecureClient;

pub use crate::state::snapshot::Snapshot;

pub use crate::utils::StatusMessage;

pub use crate::ResultMessage;

pub use self::procedures::ProcResult;

Modules

messages

Message types for [SecureClientActor]

procedures

Macros

impl_handler

Functional macro to remove boilerplate code for the implementation of the [SecureActor]. TODO Make receiver type pass as argument.

Enums

SnapshotError

StoreError

VaultError

Type Definitions

Store

Store typedef on engine::store::Cache