Module iota_stronghold::interface::comm

communication feature relevant imports are bundled here.

Re-exports

pub use crate::utils::ResultMessage;

Structs

BehaviourConfig
Configuration for initiating the P2PNetworkBehaviour.

CommunicationActor
Actor responsible for creating a [P2PNetworkBehaviour] and handling all interaction with the Swarm. For each received CommunicationRequest, a CommunicationResults is returned to the sender.

CommunicationActorConfig
The actor configuration

CommunicationActorProxy

EstablishedConnection
Information about the connection with a remote peer as maintained in the ConnectionManager.

FirewallPermission
The sum of allowed permissions. This is using the same concepts as e.g. permission values in Unix systems.

Multiaddr
Representation of a Multiaddr.

PeerId
Identifier of a peer of the network.

Enums

CommunicationRequest
Requests for the [CommunicationActor].

CommunicationResults
Returned results from the [CommunicationActor]

FirewallRule
Configure the firewall.

Keypair
Identity keypair of a node.

RelayDirection
Direction for which a relay peer is used.

RequestDirection
The direction of a [CommunicationRequest::RequestMsg] that firewall receives.

Traits

VariantPermission
The permission value for the different variants of an enum. This allows permitting specific variants of an enum while prohibiting others. In structs or unions, it should default to PermissionValue(1)