Struct iota_stronghold::interface::Stronghold

pub struct Stronghold {
    pub system: ActorSystem,
    clients: HashMap<ClientId, ActorRef<ClientMsg>>,
    target: ActorRef<ClientMsg>,
    communication_actor: Option<ActorRef<CommunicationRequest<SHRequest, ClientMsg>>>,
}

The main type for the Stronghold System. Used as the entry point for the actor model. Contains various pieces of metadata to interpret the data in the vault and store.

Fields
system: ActorSystemclients: HashMap<ClientId, ActorRef<ClientMsg>>target: ActorRef<ClientMsg>communication_actor: Option<ActorRef<CommunicationRequest<SHRequest, ClientMsg>>>
Implementations

impl Stronghold

pub fn init_stronghold_system(
system: ActorSystem,
client_path: Vec<u8>,
_options: Vec<StrongholdFlags>
) -> Self

Initializes a new instance of the system. Sets up the first client actor. Accepts a ActorSystem, the first client_path: Vec<u8> and any StrongholdFlags which pertain to the first actor.

pub async fn spawn_stronghold_actor(
&mut self,
client_path: Vec<u8>,
_options: Vec<StrongholdFlags>
) -> StatusMessage

Spawns a new set of actors for the Stronghold system. Accepts the client_path: Vec<u8> and the options: StrongholdFlags

pub async fn switch_actor_target(
&mut self,
client_path: Vec<u8>
) -> StatusMessage

Switches the actor target to another actor in the system specified by the client_path: Vec<u8>.

pub async fn write_to_vault(
&self,
location: Location,
payload: Vec<u8>,
hint: RecordHint,
_options: Vec<VaultFlags>
) -> StatusMessage

Writes data into the Stronghold. Uses the current target actor as the client and writes to the specified location of Location type. The payload must be specified as a Vec<u8> and a RecordHint can be provided. Also accepts VaultFlags for when a new Vault is created.

pub async fn write_to_store(
&self,
location: Location,
payload: Vec<u8>,
lifetime: Option<Duration>
) -> StatusMessage

Writes data into an insecure cache. This method, accepts a Location, a Vec<u8> and an optional Duration. The lifetime allows the data to be deleted after the specified duration has passed. If no lifetime is specified, the data will persist until it is manually deleted or over-written. Note: One store is mapped to one client. Can specify the same location across multiple clients.

pub async fn read_from_store(
&self,
location: Location
) -> (Vec<u8>, StatusMessage)

A method that reads from an insecure cache. This method, accepts a Location and returns the payload in the form of a (Vec<u8>, StatusMessage). If the location does not exist, an empty vector will be returned along with an error StatusMessage. Note: One store is mapped to one client. Can specify the same location across multiple clients.

pub async fn delete_from_store(&self, location: Location) -> StatusMessage

A method to delete data from an insecure cache. This method, accepts a Location and returns a StatusMessage. Note: One store is mapped to one client. Can specify the same location across multiple clients.

pub async fn delete_data(
&self,
location: Location,
should_gc: bool
) -> StatusMessage

Revokes the data from the specified location of type Location. Revoked data is not readable and can be removed from a vault with a call to garbage_collect. if the should_gc flag is set to true, this call with automatically cleanup the revoke. Otherwise, the data is just marked as revoked.

pub async fn garbage_collect(&self, vault_path: Vec<u8>) -> StatusMessage

Garbage collects any revokes in a Vault based on the given vault_path and the current target actor.

pub async fn list_hints_and_ids<V: Into<Vec<u8>>>(
&self,
vault_path: V
) -> (Vec<(RecordId, RecordHint)>, StatusMessage)

Returns a list of the available RecordId and RecordHint values in a vault by the given vault_path.

pub async fn runtime_exec(&self, control_request: Procedure) -> ProcResult

Executes a runtime command given a Procedure. Returns a ProcResult based off of the control_request specified.

pub async fn record_exists(&self, location: Location) -> bool

Checks whether a record exists in the client based off of the given Location.

pub async fn vault_exists(&self, location: Location) -> bool

checks whether a vault exists in the client.

pub async fn read_snapshot<T: Zeroize + AsRef<Vec<u8>>>(
&mut self,
client_path: Vec<u8>,
former_client_path: Option<Vec<u8>>,
keydata: &T,
filename: Option<String>,
path: Option<PathBuf>
) -> StatusMessage

Reads data from a given snapshot file. Can only read the data for a single client_path at a time. If the new actor uses a new client_path the former client path may be passed into the function call to read the data into that actor. Also requires keydata to unlock the snapshot. A filename and filepath can be specified. The Keydata should implement and use Zeroize.

pub async fn write_all_to_snapshot<T: Zeroize + AsRef<Vec<u8>>>(
&mut self,
keydata: &T,
filename: Option<String>,
path: Option<PathBuf>
) -> StatusMessage

Writes the entire state of the Stronghold into a snapshot. All Actors and their associated data will be written into the specified snapshot. Requires keydata to encrypt the snapshot and a filename and path can be specified. The Keydata should implement and use Zeroize.

pub async fn kill_stronghold(
&mut self,
client_path: Vec<u8>,
kill_actor: bool
) -> StatusMessage

Used to kill a stronghold actor or clear the cache of the given actor system based on the client_path. If kill_actor is true both the internal actor and the client actor will be killed. Otherwise, the cache of the current target actor will be cleared.

fn check_config_flags()

Unimplemented until Policies are implemented.

impl Stronghold

pub fn spawn_communication_with_keypair(
&mut self,
keypair: Keypair
) -> StatusMessage

Spawn the communication actor and swarm with a pre-existing keypair Per default, the firewall allows all outgoing, and reject all incoming requests.

pub fn spawn_communication(&mut self) -> StatusMessage

Spawn the communication actor and swarm. Per default, the firewall allows all outgoing, and reject all incoming requests.

pub fn stop_communication(&mut self)

Gracefully stop the communication actor and swarm

pub async fn start_listening(
&self,
addr: Option<Multiaddr>
) -> ResultMessage<Multiaddr>

Start listening on the swarm to the given address. If not address is provided, it will be assigned by the OS.

pub async fn stop_listening(&self) -> StatusMessage

Stop listening on the swarm.

pub async fn get_swarm_info(
&self
) -> ResultMessage<(PeerId, Vec<Multiaddr>, Vec<(PeerId, EstablishedConnection)>)>

Get the peer id, listening addresses and connection info of the local peer

pub async fn add_peer(
&self,
peer_id: PeerId,
addr: Option<Multiaddr>,
is_relay: Option<RelayDirection>
) -> ResultMessage<PeerId>

Add dial information for a remote peers. This will attempt to connect the peer directly either by the address if one is provided, or by peer id if the peer is already known e.g. from multicast DNS. If the peer is not a relay and can not be reached directly, it will be attempted to reach it via the relays, if there are any. Relays can be used to listen for incoming request, or to connect to a remote peer that can not be reached directly, and is listening to the same relay. Once the peer was successfully added, it can be used as target for operations on the remote stronghold.

pub async fn change_relay_direction(
&self,
peer_id: PeerId,
direction: RelayDirection
) -> ResultMessage<PeerId>

Set / overwrite the direction for which relay is used. RelayDirection::Dialing adds the relay to the list of relay nodes that are tried if a peer can not be reached directly. RelayDirection::Listening connect the local system with the given relay and allows that it can be reached by remote peers that use the same relay for dialing. The relay has to be added beforehand with its multi-address via the add_peer method.

pub async fn remove_relay(&self, peer_id: PeerId) -> StatusMessage

Remove a relay so that it will not be used anymore for dialing or listening.

pub async fn allow_all_requests(
&self,
peers: Vec<PeerId>,
set_default: bool
) -> StatusMessage

Allow all requests from the given peers, optionally also set default to allow all.

pub async fn allow_requests(
&self,
peers: Vec<PeerId>,
change_default: bool,
requests: Vec<SHRequestPermission>
) -> StatusMessage

Change or add rules in the firewall to allow the given requests for the peers, optionally also change the default rule to allow it. The SHRequestPermission copy the SHRequest with Unit-type variants with individual permission, e.g.

use iota_stronghold::SHRequestPermission;

let permissions = vec![SHRequestPermission::CheckVault, SHRequestPermission::CheckRecord];

Existing permissions for other SHRequestPermissions will not be changed by this. If no rule has been set for a given peer, the default rule will be used as basis.

pub async fn reject_requests(
&self,
peers: Vec<PeerId>,
change_default: bool,
requests: Vec<SHRequestPermission>
) -> StatusMessage

Change or add rules in the firewall to reject the given requests from the peers, optionally also remove the permission from the default rule. The SHRequestPermission copy the SHRequest with Unit-type variants with individual permission, e.g.

use iota_stronghold::SHRequestPermission;

let permissions = vec![SHRequestPermission::CheckVault, SHRequestPermission::CheckRecord];

Existing permissions for other SHRequestPermissions will not be changed by this. If no rule has been set for a given peer, the default rule will be used as basis.

pub async fn reject_all_requests(
&self,
peers: Vec<PeerId>,
set_default: bool
) -> StatusMessage

Configure the firewall to reject all requests from the given peers, optionally also set default rule to reject all.

pub async fn remove_firewall_rules(&self, peers: Vec<PeerId>) -> StatusMessage

Remove peer specific rules from the firewall configuration.

pub async fn write_remote_vault(
&self,
peer_id: PeerId,
location: Location,
payload: Vec<u8>,
hint: RecordHint,
_options: Vec<VaultFlags>
) -> StatusMessage

Write to the vault of a remote Stronghold. It is required that the peer has successfully been added with the add_peer method.

pub async fn write_to_remote_store(
&self,
peer_id: PeerId,
location: Location,
payload: Vec<u8>,
lifetime: Option<Duration>
) -> StatusMessage

Write to the store of a remote Stronghold. It is required that the peer has successfully been added with the add_peer method.

pub async fn read_from_remote_store(
&self,
peer_id: PeerId,
location: Location
) -> (Vec<u8>, StatusMessage)

Read from the store of a remote Stronghold. It is required that the peer has successfully been added with the add_peer method.

pub async fn list_remote_hints_and_ids<V: Into<Vec<u8>>>(
&self,
peer_id: PeerId,
vault_path: V
) -> (Vec<(RecordId, RecordHint)>, StatusMessage)

Returns a list of the available records and their RecordHint values of a remote vault. It is required that the peer has successfully been added with the add_peer method.

pub async fn remote_runtime_exec(
&self,
peer_id: PeerId,
control_request: Procedure
) -> ProcResult

Executes a runtime command at a remote Stronghold. It is required that the peer has successfully been added with the add_peer method.

async fn ask_remote(
&self,
peer_id: PeerId,
request: SHRequest
) -> Result<SHResults, String>

async fn configure_firewall(&self, rule: FirewallRule) -> StatusMessage

async fn ask_communication_actor(
&self,
request: CommunicationRequest<SHRequest, ClientMsg>
) -> Result<CommunicationResults<SHResults>, String>

pub fn keep_alive<F>(&self, callback: Option<F>) where
F: FnOnce() -> Result<(), Box<dyn Error>>,

Keeps stronghold in a running state. This call is blocking.

This function accepts an optional function for more control over how long stronghold shall block.