Struct iota_stronghold::state::bucket::Bucket

pub struct Bucket {
    cache: HashMap<VaultId, Vec<ReadResult>>,
}

A Bucket cache of the Data for stronghold. Contains a HashMap<Key<P>, Option<DBView<P>>> pairing the vault Key<P> and the vault DBView<P> together. Also contains a HashMap<Key<P>, Vec<ReadResult>> which pairs the backing data with the associated Key<P>.

Fields
cache: HashMap<VaultId, Vec<ReadResult>>
Implementations

impl Bucket

pub fn new() -> Self

Creates a new Bucket.

pub fn create_and_init_vault<P: BoxProvider>(
&mut self,
id: VaultId,
key: Key<P>,
rid: RecordId
) -> RecordId

Creates and initializes a new Vault given a Key<P>. Returns a tuple of (Key<P>, RecordId). The returned Key<P> is the Key associated with the Vault and the RecordId is the ID for its first record.

pub fn read_data<P: BoxProvider>(
&mut self,
id: VaultId,
key: Key<P>,
rid: RecordId
) -> Vec<u8>

Reads data from a Record in the Vault given a RecordId. Returns the data as a Vec<u8> of utf8 bytes.

pub fn record_exists_in_vault<P: BoxProvider>(
&mut self,
id: VaultId,
key: Key<P>,
rid: RecordId
) -> bool

pub fn init_record<P: BoxProvider>(
&mut self,
id: VaultId,
key: Key<P>,
rid: RecordId
) -> RecordId

Initializes a new Record in the Vault based on the inserted Key<P>. Returns a RecordId for the new Record.

pub fn write_payload<P: BoxProvider>(
&mut self,
id: VaultId,
key: Key<P>,
rid: RecordId,
payload: Vec<u8>,
hint: RecordHint
)

Writes a payload of Vec<u8> and a RecordHint into a Record. Record is specified with the inserted RecordId and the Key<P>

pub fn revoke_data<P: BoxProvider>(
&mut self,
id: VaultId,
key: Key<P>,
rid: RecordId
)

Marks a record for deletion based on a given Key<P> and RecordId

pub fn garbage_collect<P: BoxProvider>(&mut self, id: VaultId, key: Key<P>)

Garbage Collects any deletion marked Records in the given Vault. Accepts a Key<P>

pub fn list_ids<P: BoxProvider>(
&mut self,
id: VaultId,
key: Key<P>
) -> Vec<(RecordId, RecordHint)>

Lists the RecordIds and RecordHints for a given Vault. Accepts a Key<P>

pub fn repopulate_data(&mut self, cache: HashMap<VaultId, Vec<ReadResult>>)

Repopulates the data in the Bucket given a Vec of state from a snapshot.

pub fn get_data(&mut self) -> HashMap<VaultId, Vec<ReadResult>>

pub fn clear_cache(&mut self)

fn take<P: BoxProvider>(
&mut self,
id: VaultId,
key: Key<P>,
f: impl FnOnce(DBView<P>, Vec<ReadResult>) -> Vec<ReadResult>
)

Exposes the DBView of the current vault and the cache layer to allow transactions to occur.

fn get_reads(&mut self, id: VaultId) -> Vec<ReadResult>

fn insert_reads(&mut self, id: VaultId, reads: Vec<ReadResult>)