Struct iota_stronghold::state::key_store::KeyStore

pub struct KeyStore<P: BoxProvider + Clone + Send + Sync + 'static> {
    store: HashMap<VaultId, Key<P>>,
}

Fields
store: HashMap<VaultId, Key<P>>
Implementations

impl<P: BoxProvider + Clone + Send + Sync + 'static> KeyStore<P>

pub fn new() -> Self

Creates a new KeyStore.

pub fn get_key(&mut self, id: VaultId) -> Option<Key<P>>

Gets the key from the KeyStore and removes it. Returns an Option<Key<P>>

pub fn vault_exists(&self, id: VaultId) -> bool

Checks to see if the vault exists.

pub fn create_key(&mut self, id: VaultId) -> Key<P>

Returns an existing key for the id or creates one.

pub fn insert_key(&mut self, id: VaultId, key: Key<P>) -> &Key<P>

Inserts a key into the KeyStore by VaultId. If the VaultId already exists, it just returns the existing [&Key<P>]

pub fn rebuild_keystore(&mut self, keys: HashMap<VaultId, Key<P>>)

Rebuilds the KeyStore while throwing out any existing [VauldId], Key<P> pairs. Accepts a Vec<Key<P>> and returns the a Vec<VaultId>; primarily used to repopulate the state from a snapshot.

pub fn get_data(&mut self) -> HashMap<VaultId, Key<P>>

Gets the state data in a hashmap format for the snapshot.

pub fn clear_keys(&mut self)

Clear the key store.