Struct iota_stronghold::state::secure::SecureClient

pub struct SecureClient {
    pub(crate) keystore: KeyStore<Provider>,
    pub(crate) db: DbView<Provider>,
    pub client_id: ClientId,
    pub vaults: HashSet<VaultId>,
    pub store: Cache<Vec<u8>, Vec<u8>>,
}

Fields
keystore: KeyStore<Provider>db: DbView<Provider>client_id: ClientIdvaults: HashSet<VaultId>store: Cache<Vec<u8>, Vec<u8>>
Implementations

impl SecureClient

pub fn new(client_id: ClientId) -> Self

Creates a new Client given a ClientID and ChannelRef<SHResults>

pub fn write_to_store(
&mut self,
key: Vec<u8>,
data: Vec<u8>,
lifetime: Option<Duration>
) -> Option<Vec<u8>>

Write unencrypted data to the store. Returns None if the key didn’t already exist and [Some(Vec<u8>)] if the key was updated.

pub fn read_from_store(&mut self, key: Vec<u8>) -> Option<Vec<u8>>

Attempts to read the data from the store. Returns [Some(Vec<u8>)] if the key exists and None if it doesn’t.

pub fn store_delete_item(&mut self, key: Vec<u8>)

Deletes an item from the store by the given key.

pub fn store_key_exists(&mut self, key: Vec<u8>) -> bool

Checks to see if the key exists in the store.

pub fn set_client_id(&mut self, client_id: ClientId)

Sets the client id to swap from one client to another.

pub fn add_new_vault(&mut self, vid: VaultId)

Adds a new vault to the client. If the VaultId already exists, will just use that existing Vault.

pub fn clear_cache(&mut self) -> Option<()>

Empty the Client Cache.

pub fn rebuild_cache(
&mut self,
id: ClientId,
vaults: HashSet<VaultId>,
store: Cache<Vec<u8>, Vec<u8>>
)

Rebuilds the cache using the parameters.

pub fn resolve_location<L: AsRef<Location>>(&self, l: L) -> (VaultId, RecordId)

Resolves a location to a VaultId and a RecordId

pub fn derive_vault_id<P: AsRef<Vec<u8>>>(&self, path: P) -> VaultId

Gets the VaultId from a specified path.

pub fn derive_record_id<P: AsRef<Vec<u8>>>(
&self,
vault_path: P,
ctr: usize
) -> RecordId

Derives the counter RecordId from the given vault path and the counter value.

pub fn get_client_str(&self) -> String

Gets the client string.

pub fn vault_exist(&self, vid: VaultId) -> Option<&VaultId>

Checks to see if the vault exists in the client.

pub fn get_index_from_record_id<P: AsRef<Vec<u8>>>(
&self,
vault_path: P,
record_id: RecordId
) -> usize

Gets the current index of a record if its a counter.