Struct iota_stronghold::state::snapshot::Snapshot

pub struct Snapshot {
    pub state: SnapshotState,
}

Wrapper for the SnapshotState data structure.

Fields
state: SnapshotState
Implementations

impl Snapshot

pub fn new(state: SnapshotState) -> Self

Creates a new Snapshot from a buffer of SnapshotState state.

pub fn get_state(
&mut self,
id: ClientId
) -> (HashMap<VaultId, PKey<Provider>>, DbView<Provider>, Cache<Vec<u8>, Vec<u8>>)

Gets the state component parts as a tuple.

pub fn has_data(&self, cid: ClientId) -> bool

Checks to see if the ClientId exists in the snapshot hashmap.

pub fn read_from_snapshot(
name: Option<&str>,
path: Option<&Path>,
key: Key
) -> Result<Self>

Reads state from the specified named snapshot or the specified path TODO: Add associated data.

pub fn write_to_snapshot(
&self,
name: Option<&str>,
path: Option<&Path>,
key: Key
) -> Result<()>

Writes state to the specified named snapshot or the specified path TODO: Add associated data.