Struct iota_stronghold::ActorSystem

pub struct ActorSystem {
    proto: Arc<ProtoSystem>,
    sys_actors: Option<SysActors>,
    log: LoggingSystem,
    debug: bool,
    pub exec: ThreadPool,
    pub timer: Sender<Job>,
    pub sys_channels: Option<SysChannels>,
    pub(crate) provider: Provider,
}

The actor runtime and common services coordinator

The ActorSystem provides a runtime on which actors are executed. It also provides common services such as channels and scheduling. The ActorSystem is the heart of a Riker application, starting several threads when it is created. Create only one instance of ActorSystem per application.

Fields
proto: Arc<ProtoSystem>sys_actors: Option<SysActors>log: LoggingSystemdebug: boolexec: ThreadPooltimer: Sender<Job>sys_channels: Option<SysChannels>provider: Provider
Implementations

impl ActorSystem

pub fn new() -> Result<ActorSystem, SystemError>

Create a new ActorSystem instance

Requires a type that implements the Model trait.

pub fn with_name(name: &str) -> Result<ActorSystem, SystemError>

Create a new ActorSystem instance with provided name

Requires a type that implements the Model trait.

pub fn with_config(name: &str, cfg: Config) -> Result<ActorSystem, SystemError>

Create a new ActorSystem instance bypassing default config behavior

pub fn start_date(&self) -> &DateTime<Utc>

Returns the system start date

pub fn uptime(&self) -> u64

Returns the number of seconds since the system started

pub fn host(&self) -> Arc<str>

Returns the hostname used when the system started

The host is used in actor addressing.

Currently not used, but will be once system clustering is introduced.

pub fn id(&self) -> Uuid

Returns the UUID assigned to the system

pub fn name(&self) -> String

Returns the name of the system

pub fn print_tree(&self)

pub fn user_root(&self) -> &BasicActorRef

Returns the user root actor reference

pub fn sys_root(&self) -> &BasicActorRef

Returns the system root actor reference

pub fn temp_root(&self) -> &BasicActorRef

Reutrns the temp root actor reference

pub fn sys_events(&self) -> &ActorRef<ChannelMsg<SystemEvent>>

Returns a reference to the system events channel

pub fn dead_letters(&self) -> &ActorRef<ChannelMsg<DeadLetter>>

Returns a reference to the dead letters channel

pub fn publish_event(&self, evt: SystemEvent)

pub fn config(&self) -> &Config

Returns the Config used by the system

pub fn sys_actor_of_props<A>(
&self,
name: &str,
props: Arc<Mutex<dyn ActorProducer<Actor = A> + 'static>>
) -> Result<ActorRef<<A as Actor>::Msg>, CreateError> where
A: Actor,

Create an actor under the system root

pub fn sys_actor_of<A>(
&self,
name: &str
) -> Result<ActorRef<<A as Actor>::Msg>, CreateError> where
A: ActorFactory,

pub fn sys_actor_of_args<A, Args>(
&self,
name: &str,
args: Args
) -> Result<ActorRef<<A as Actor>::Msg>, CreateError> where
A: ActorFactoryArgs<Args>,
Args: ActorArgs,

pub fn log(&self) -> LoggingSystem

pub fn shutdown(&self) -> Receiver<()>

Shutdown the actor system

Attempts a graceful shutdown of the system and all actors. Actors will receive a stop message, executing actor.post_stop.

Does not block. Returns a future which is completed when all actors have successfully stopped.