Trait iota_stronghold::utils::ids::LoadFromPath

pub trait LoadFromPath: Sized {
    fn load_from_path(data: &[u8], path: &[u8]) -> Result<Self>;
}

A trait that allows a datatype to load and setup its internal data through the use of a path and some data.

Required methods

fn load_from_path(data: &[u8], path: &[u8]) -> Result<Self>

Load from some data and a path.