Enum iota_stronghold::utils::types::Location

pub enum Location {
    Generic {
        vault_path: Vec<u8>,
        record_path: Vec<u8>,
    },
    Counter {
        vault_path: Vec<u8>,
        counter: usize,
    },
}

A Location type used to specify where in the Stronghold a piece of data should be stored. A generic location specifies a non-versioned location while a counter location specifies a versioned location. The Counter location can be used to get the head of the version chain by passing in None as the counter index. Otherwise, counter records are referenced through their associated index. On Read, the None location is the latest record in the version chain while on Write, the None location is the next record in the version chain.

Variants

Generic

Fields of Generic

vault_path: Vec<u8>record_path: Vec<u8>

Counter

Fields of Counter

vault_path: Vec<u8>counter: usize

Implementations

impl Location

pub fn vault_path(&self) -> &[u8]

Gets the vault_path from the Location.

pub fn generic<V: Into<Vec<u8>>, R: Into<Vec<u8>>>(
vault_path: V,
record_path: R
) -> Self

Creates a generic location from types that implement Into<Vec<u8>>.

pub fn counter<V: Into<Vec<u8>>, C: Into<usize>>(
vault_path: V,
counter: C
) -> Self

Creates a counter location from a type that implements Into<Vec<u8>> and a counter type that implements Into<usize>

pub fn increment_counter(self) -> Self

Helper method used to increment counter locations.

pub const fn const_generic(vault_path: Vec<u8>, record_path: Vec<u8>) -> Self

Used to generate a constant generic location.

pub const fn const_counter(vault_path: Vec<u8>, counter: usize) -> Self

used to generate a constant counter location.