Crate iota_stronghold

Modules

actors

interface

internals

state

utils

Macros

line_error
TODO: Should be replaced with proper errors.

Structs

Provider
An implementation of the Vault's `BoxProvider type. Used to encrypt and decrypt the data in this Stronghold.

RecordHint
a record hint. Used as a hint to what this data is.

Stronghold
The main type for the Stronghold System. Used as the entry point for the actor model. Contains various pieces of metadata to interpret the data in the vault and store.

Enums

Error

Location
A Location type used to specify where in the Stronghold a piece of data should be stored. A generic location specifies a non-versioned location while a counter location specifies a versioned location. The Counter location can be used to get the head of the version chain by passing in None as the counter index. Otherwise, counter records are referenced through their associated index. On Read, the None location is the latest record in the version chain while on Write, the None location is the next record in the version chain.

ProcResult
A Procedure return result type. Contains the different return values for the Procedure type calls used with Stronghold.runtime_exec(...).

Procedure
Procedure type used to call to the runtime via Strongnhold.runtime_exec(...).

ResultMessage
Return value used for Actor Messages. Can specify an Error or an Ok result.

SLIP10DeriveInput
SLIP10DeriveInput type used to specify a Seed location or a Key location for the SLIP10Derive procedure.

StrongholdFlags
Policy options for modifying an entire Stronghold. Must be specified on creation.

VaultFlags
Policy options for for a specific vault. Must be specified on creation.

Functions

home_dir
Get the preferred Stronghold home directory

naive_kdf
a wrapper around the HMAC_SHA256 function used to derive a hash from a given password.

snapshot_dir
Get the preferred snapshot directory

Type Definitions

Key

Result

StatusMessage
A type alias for the empty ResultMessage<()> type.