Struct iota_stronghold::interface::Stronghold

pub struct Stronghold {
    registry: Addr<Registry>,
    target: Addr<SecureClient>,
    network_actor: Option<Addr<NetworkActor>>,
}

The main type for the Stronghold System. Used as the entry point for the actor model. Contains various pieces of metadata to interpret the data in the vault and store.

Fields
registry: Addr<Registry>target: Addr<SecureClient>network_actor: Option<Addr<NetworkActor>>
Implementations

impl Stronghold

pub async fn init_stronghold_system(
client_path: Vec<u8>,
_options: Vec<StrongholdFlags>
) -> Result<Self, Error>

Initializes a new instance of the system asynchronously. Sets up the first client actor. Accepts the first client_path: Vec<u8> and any StrongholdFlags which pertain to the first actor. The [actix::SystemRunner] is not being used directly by stronghold, and must be initialized externally.

pub async fn spawn_stronghold_actor(
&mut self,
client_path: Vec<u8>,
_options: Vec<StrongholdFlags>
) -> StatusMessage

Spawns a new set of actors for the Stronghold system. Accepts the client_path: Vec<u8> and the options: StrongholdFlags

pub async fn switch_actor_target(
&mut self,
client_path: Vec<u8>
) -> StatusMessage

Switches the actor target to another actor in the system specified by the client_path: Vec<u8>.

pub async fn write_to_vault(
&self,
location: Location,
payload: Vec<u8>,
hint: RecordHint,
_options: Vec<VaultFlags>
) -> StatusMessage

Writes data into the Stronghold. Uses the current target actor as the client and writes to the specified location of Location type. The payload must be specified as a Vec<u8> and a RecordHint can be provided. Also accepts VaultFlags for when a new Vault is created.

pub async fn write_to_store(
&self,
location: Location,
payload: Vec<u8>,
lifetime: Option<Duration>
) -> StatusMessage

Writes data into an insecure cache. This method, accepts a Location, a Vec<u8> and an optional Duration. The lifetime allows the data to be deleted after the specified duration has passed. If no lifetime is specified, the data will persist until it is manually deleted or over-written. Note: One store is mapped to one client. Can specify the same location across multiple clients.

pub async fn read_from_store(
&self,
location: Location
) -> (Vec<u8>, StatusMessage)

A method that reads from an insecure cache. This method, accepts a Location and returns the payload in the form of a (Vec<u8>, StatusMessage). If the location does not exist, an empty vector will be returned along with an error StatusMessage. Note: One store is mapped to one client. Can specify the same location across multiple clients.

pub async fn delete_from_store(&self, location: Location) -> StatusMessage

A method to delete data from an insecure cache. This method, accepts a Location and returns a StatusMessage. Note: One store is mapped to one client. Can specify the same location across multiple clients.

pub async fn delete_data(
&self,
location: Location,
should_gc: bool
) -> StatusMessage

Revokes the data from the specified location of type Location. Revoked data is not readable and can be removed from a vault with a call to garbage_collect. if the should_gc flag is set to true, this call with automatically cleanup the revoke. Otherwise, the data is just marked as revoked.

pub async fn garbage_collect(&self, vault_path: Vec<u8>) -> StatusMessage

Garbage collects any revokes in a Vault based on the given vault_path and the current target actor.

pub async fn list_hints_and_ids<V: Into<Vec<u8>>>(
&self,
vault_path: V
) -> (Vec<(RecordId, RecordHint)>, StatusMessage)

Returns a list of the available RecordId and RecordHint values in a vault by the given vault_path.

pub async fn runtime_exec(&self, control_request: Procedure) -> ProcResult

Executes a runtime command given a Procedure. Returns a ProcResult based off of the control_request specified.

pub async fn record_exists(&self, location: Location) -> bool

Checks whether a record exists in the client based off of the given Location.

pub async fn vault_exists(&self, location: Location) -> bool

checks whether a vault exists in the client.

pub async fn read_snapshot<T: Zeroize + AsRef<Vec<u8>>>(
&mut self,
client_path: Vec<u8>,
former_client_path: Option<Vec<u8>>,
keydata: &T,
filename: Option<String>,
path: Option<PathBuf>
) -> StatusMessage

Reads data from a given snapshot file. Can only read the data for a single client_path at a time. If the new actor uses a new client_path the former client path may be passed into the function call to read the data into that actor. Also requires keydata to unlock the snapshot. A filename and filepath can be specified. The Keydata should implement and use Zeroize.

pub async fn write_all_to_snapshot<T: Zeroize + AsRef<Vec<u8>>>(
&mut self,
keydata: &T,
filename: Option<String>,
path: Option<PathBuf>
) -> StatusMessage

Writes the entire state of the Stronghold into a snapshot. All Actors and their associated data will be written into the specified snapshot. Requires keydata to encrypt the snapshot and a filename and path can be specified. The Keydata should implement and use Zeroize.

pub async fn kill_stronghold(
&mut self,
client_path: Vec<u8>,
kill_actor: bool
) -> StatusMessage

Used to kill a stronghold actor or clear the cache of the given actor system based on the client_path. If kill_actor is true, the actor will be removed from the system. Otherwise, the cache of the current target actor will be cleared.

fn check_config_flags()

Unimplemented until Policies are implemented.

impl Stronghold

pub async fn spawn_p2p(
&mut self,
firewall_rule: Rule<ShRequest>,
network_config: NetworkConfig
) -> StatusMessage

Spawn the p2p-network actor and swarm.

pub fn stop_p2p(&mut self)

Gracefully stop the network actor and swarm.

pub async fn start_listening(
&self,
address: Option<Multiaddr>
) -> ResultMessage<Multiaddr>

Start listening on the swarm to the given address. If not address is provided, it will be assigned by the OS.

pub async fn stop_listening(&self) -> StatusMessage

Stop listening on the swarm.

pub async fn get_swarm_info(&self) -> ResultMessage<SwarmInfo>

Get the peer id, listening addresses and connection info of the local peer

pub async fn add_peer(
&self,
peer: PeerId,
address: Option<Multiaddr>,
is_listening_relay: bool,
is_dialing_relay: bool
) -> StatusMessage

Add dial information for a remote peers. This will attempt to connect the peer directly either by the address if one is provided, or by peer id if the peer is already known e.g. from multicast DNS. If the peer is not a relay and can not be reached directly, it will be attempted to reach it via the relays, if there are any. Relays can be used to listen for incoming request, or to connect to a remote peer that can not be reached directly, and is listening to the same relay. Once the peer was successfully added, it can be used as target for operations on the remote stronghold.

pub async fn remove_relay(
&self,
relay: PeerId,
rm_listening_relay: bool,
rm_dialing_relay: bool
) -> StatusMessage

Remove a peer from the list of peers used for dialing, and / or stop listening with the relay.

pub async fn set_firewall_rule(
&self,
rule: Rule<ShRequest>,
peers: Vec<PeerId>,
set_default: bool
) -> StatusMessage

Change the firewall rule for specific peers, optionally also set it as the default rule, which applies if there are no specific rules for a peer. All inbound requests from the peers that this rule applies to, will be approved/ rejected based on this rule.

pub async fn remove_firewall_rules(&self, peers: Vec<PeerId>) -> StatusMessage

Remove peer specific rules from the firewall configuration.

pub async fn write_remote_vault(
&self,
peer: PeerId,
location: Location,
payload: Vec<u8>,
hint: RecordHint,
_options: Vec<VaultFlags>
) -> StatusMessage

Write to the vault of a remote Stronghold. It is required that the peer has successfully been added with the add_peer method.

pub async fn write_to_remote_store(
&self,
peer: PeerId,
location: Location,
payload: Vec<u8>,
lifetime: Option<Duration>
) -> StatusMessage

Write to the store of a remote Stronghold. It is required that the peer has successfully been added with the add_peer method.

pub async fn read_from_remote_store(
&self,
peer: PeerId,
location: Location
) -> ResultMessage<Vec<u8>>

Read from the store of a remote Stronghold. It is required that the peer has successfully been added with the add_peer method.

pub async fn list_remote_hints_and_ids<V: Into<Vec<u8>>>(
&self,
peer: PeerId,
vault_path: V
) -> ResultMessage<Vec<(RecordId, RecordHint)>>

Returns a list of the available records and their RecordHint values of a remote vault. It is required that the peer has successfully been added with the add_peer method.

pub async fn remote_runtime_exec(
&self,
peer: PeerId,
control_request: Procedure
) -> ResultMessage<ProcResult>

Executes a runtime command at a remote Stronghold. It is required that the peer has successfully been added with the add_peer method.